refactor(#7): 패키지명 오류 수정
